refactor: Scheduler 객체 분리
