+++ app/src/main/java/woowacourse/movie/domain/DateScheduler.kt
@@ -1,0 +1,34 @@
+package woowacourse.movie.domain
+
+import java.time.LocalDate
+import java.time.temporal.ChronoUnit
+
+class DateScheduler {
+    fun reservableDates(
+        screeningDate: ScreeningDate,
+        currentDate: LocalDate,
+    ): List<LocalDate> {
+        val startDate = getStartDate(screeningDate.startDate, currentDate)
+        return getDatesBetween(startDate, screeningDate.endDate)
+    }
+
+    fun getStartDate(
+        movieStartDate: LocalDate,
+        currentDate: LocalDate,
+    ): LocalDate {
+        return if (movieStartDate < currentDate) {
+            currentDate
+        } else {
+            movieStartDate
+        }
+    }
+
+    private fun getDatesBetween(
+        startDate: LocalDate,
+        endDate: LocalDate,
+    ): List<LocalDate> {
+        val dayDiff = ChronoUnit.DAYS.between(startDate, endDate)
+
+        return (0..dayDiff).map { startDate.plusDays(it) }
+    }
+}
